+++ src/controllers/admin/jobDesignationController.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 import { injectable, inject } from "tsyringe";
 import { IjobDesignationService } from "../../interfaces/Iservices/IadminService/IjobDesignationService";
 import { AddDesignationResponseDTO } from "../../interfaces/DTO/IServices/Iadminservices.dto/jobDesignationService.dto";
-import { IjobDesignationController } from "../../interfaces/Icontrollers/IjobDesignationController";
+import { IjobDesignationController } from "../../interfaces/Icontrollers/Iadmincontrollers/IjobDesignationController";
 import { HTTP_STATUS } from "../../utils/httpStatus";
 
 @injectable()
